extractor/requirements.txt
mysqlclient==2.2.7
numpy==2.2.5 
pandas==2.2.3 
python-dateutil==2.9.0.post0 
python-dotenv==1.1.0 
pytz==2025.2 
six==1.17.0 
tenacity==9.1.2 
tzdata==2025.2